git-svn-id: https://bsisvn.mayo.edu/main/projects/g/Genome_GPS/trunk@5883 8f329c14-d232-4f17-8684-7ee34322b8dc

+++ fastq.sh
@@ -40,7 +40,13 @@ else
         $fastqc_path/fastqc -o $fastqc_dir/ $output/$read
 		rm $fastqc_dir/${file1}_fastqc.zip
     else
-        ln -s $FOLDER_FASTQC/${file1}_fastqc $fastqc_dir/
+        if [ -f $FOLDER_FASTQC/${file1}_fastqc ]
+		then
+			ln -s $FOLDER_FASTQC/${file1}_fastqc $fastqc_dir/
+		else
+			echo "ERROR : $FOLDER_FASTQC mentioned doesn't have any fastqc results"
+			exit 1;
+		fi	
     fi		
     echo `date`
 fi
